requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
celery>=5.3
redis>=4.6
watchdog>=3.0
pandas>=2.0
openpyxl>=3.1
python-decouple>=3.8
psycopg2-binary>=2.9
django-cors-headers>=4.3
